Deployments/Django/DjangoMLTemplate/requirements.txt
pandas==1.0.3
django-ninja
django-rest-framework
django-extensions
django-htmx
bokeh
requests
lxml